skills/skill-optimizer/requirements.txt
# ==============================================
# Skill Optimizer Dependencies
# ==============================================

# Core LLM Integration
langchain-openai>=0.1.0
langchain-core>=0.1.0
langchain>=0.1.0

# API Clients
httpx>=0.24.0
requests>=2.31.0

# Environment Configuration
python-dotenv>=1.0.0

# Optional: Langfuse for Trace Recording
langfuse>=2.0.0

# Python >= 3.11 is recommended
